requirements.txt
<<<<<<< HEAD
# git+https://github.com/openai/CLIP.git
easydict==1.9
=======
easydict==1.9
faiss==1.5.3
>>>>>>> 7572dc5671b7a743af4c3aead34299d8b859cca9
faiss_gpu==1.7.2
ftfy==6.0.3
matplotlib==3.3.3
nltk==3.6.7
numpy==1.19.4
<<<<<<< HEAD
pandas==1.1.5
Pillow==8.4.0
PyYAML==6.0
regex==2022.6.2
# scikit_learn==1.1.2
scipy==1.5.4
setuptools==50.3.2
termcolor==1.1.0
tqdm==4.52.0
# torch==1.9.0
# torchvision==0.10.0
# torchaudio==0.9.0 
=======
opencv_python==4.4.0.46
pandas==1.1.5
Pillow==9.2.0
PyYAML==6.0
regex==2022.6.2
scikit_learn==1.1.2
scipy==1.5.4
setuptools==50.3.2
tensorboard==2.9.1
termcolor==1.1.0
torch==1.9.1
torchvision==0.11.2
tqdm==4.52.0
>>>>>>> 7572dc5671b7a743af4c3aead34299d8b859cca9
